backend/pyproject.toml
[tool.poetry]
name = "code-review-assistant"
version = "0.1.0"
description = ""
authors = ["Harshyara"]
package-mode=false

[tool.poetry.dependencies] 
python = "^3.13"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
google-auth = "^2.38.0"
requests = "^2.32.3"
torch = "^2.6.0"
transformers = "^4.50.0"
accelerate = "^1.5.2"
dotenv = "^0.9.9"
jsonify = "^0.5"
pydantic-settings = "^2.8.1"
redis = "^5.2.1"
pygithub = "^2.6.1"
ngrok = "^1.4.0"
google-auth-oauthlib = "^1.2.2"
jose = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
